* Add a sub-tree option to .tree()
This allows us to watch a sub-tree.

This is useful for cases when there's an existing class for a directory
(i.e., a subclass of mtDir) but we want to hook into the directory itself,
which clashes with the idea that the root must be a mtRoot.

Thus, instead of
	.tree('/foo/bar/:thing', types=t)
use
	tr = EtcTypes()
	tr.step(':thing', dest=t)
	.tree('/foo/bar',sub=':thing', types=tr)